[iOS] refactor: CellHandler 분리
